fix: UserService fix
- removed UserServiceImpl useless class
- modified UserService to provides the methods itself
- modified files to use the correct UserService
